@@ -1,57 +1,56 @@
 {
   "name": "client",
-  "version": "0.0.0",
+  "version": "1.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
+    "start": "ng serve --proxy-config src/proxy.conf.json",
+    "build": "ng build --configuration production --output-hashing none",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^9.1.9",
-    "@angular/cdk": "^9.2.4",
-    "@angular/common": "^9.1.9",
-    "@angular/compiler": "^9.1.9",
-    "@angular/core": "^9.1.9",
-    "@angular/forms": "^9.1.9",
-    "@angular/localize": "^9.1.12",
-    "@angular/platform-browser": "^9.1.9",
-    "@angular/platform-browser-dynamic": "^9.1.9",
-    "@angular/router": "^9.1.9",
-    "@swimlane/ngx-charts": "^14.0.0",
-    "@swimlane/ngx-datatable": "^17.0.0",
-    "bootstrap": "^4.5.0",
-    "chokidar": "^3.4.2",
-    "jquery": "^3.5.1",
-    "ngx-cookie-service": "^3.0.4",
+    "@angular/animations": "^13.0.3",
+    "@angular/common": "^13.0.3",
+    "@angular/compiler": "^13.0.3",
+    "@angular/core": "^13.0.3",
+    "@angular/forms": "^13.0.3",
+    "@angular/localize": "^13.0.3",
+    "@angular/platform-browser": "^13.0.3",
+    "@angular/platform-browser-dynamic": "^13.0.3",
+    "@angular/router": "^13.0.3",
+    "@angular/cdk": "^13.0.4",
+    "@swimlane/ngx-charts": "^19.1.0",
+    "@swimlane/ngx-datatable": "^20.0.0",
+    "bootstrap": "^5.1.3",
+    "chokidar": "^3.5.2",
+    "jquery": "^3.6.0",
+    "ngx-cookie-service": "^13.0.1",
     "ngx-filter-pipe": "^2.1.2",
-    "popper.js": "^1.16.1",
-    "rxjs": "^6.5.5",
-    "tslib": "^1.10.0",
-    "zone.js": "^0.10.3"
+    "@popperjs/core": "^2.11.0",
+    "rxjs": "^6.6.3",
+    "tslib": "^2.3.1",
+    "zone.js": "^0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.901.9",
-    "@angular/cli": "^9.1.7",
-    "@angular/compiler-cli": "^9.1.9",
-    "@angular/language-service": "^9.1.9",
-    "@types/jasmine": "^3.5.10",
-    "@types/jasminewd2": "^2.0.8",
-    "@types/node": "^14.0.5",
-    "codelyzer": "^5.2.2",
-    "jasmine-core": "^3.5.0",
-    "jasmine-spec-reporter": "^5.0.2",
-    "karma": "^5.0.9",
+    "@angular-devkit/build-angular": "^13.0.4",
+    "@angular/cli": "^13.0.4",
+    "@angular/compiler-cli": "^13.0.3",
+    "@angular/language-service": "^13.0.3",
+    "@types/jasmine": "^3.10.2",
+    "@types/jasminewd2": "^2.0.10",
+    "@types/node": "^16.11.12",
+    "jasmine-core": "^3.10.1",
+    "jasmine-spec-reporter": "^7.0.0",
+    "karma": "^6.3.9",
     "karma-chrome-launcher": "^3.1.0",
-    "karma-coverage-istanbul-reporter": "^3.0.2",
-    "karma-jasmine": "^3.2.0",
-    "karma-jasmine-html-reporter": "^1.5.4",
+    "karma-coverage-istanbul-reporter": "^3.0.3",
+    "karma-jasmine": "^4.0.1",
+    "karma-jasmine-html-reporter": "^1.7.0",
     "protractor": "^7.0.0",
-    "ts-node": "^8.10.1",
-    "tslint": "^6.1.2",
-    "typescript": "^3.8.3"
+    "ts-node": "^10.4.0",
+    "eslint": "^7.32.0",
+    "typescript": "~4.4.3"
   }
 }
