@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import {HttpClient, HttpHeaders} from '@angular/common/http';
-import {CookieService} from 'ngx-cookie-service';
+import {AuthService} from './services/auth.service';
+import {Router} from "@angular/router";
 
 @Component({
   selector: 'app-root',
@@ -9,5 +9,29 @@ import {CookieService} from 'ngx-cookie-service';
 })
 export class AppComponent implements OnInit {
 
-  ngOnInit() { }
+  isLoaded: boolean;
+  isAuthenticated: boolean;
+  isError: boolean;
+
+  constructor(
+    private router: Router,
+    private authService: AuthService
+  ) {
+    this.isLoaded = false;
+    this.isAuthenticated = false;
+    this.isError = false;
+  }
+
+  ngOnInit() {
+    this.authService.verify()
+      .subscribe(data => {
+        this.isAuthenticated = data === '/';
+        this.isLoaded = true;
+      },
+      error => {
+        this.isError = true;
+        this.router.navigate(['/error', {error}]).then();
+        this.isLoaded = true;
+      });
+  }
 }
