@@ -1,7 +1,7 @@
 import {Graph} from './graph.model';
 
 export interface Data {
-  key: number;
+  key: any;
   value: number;
   graph: Graph;
 }
