@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, UrlTree, Router } from '@angular/router';
-import { Observable } from 'rxjs';
+import {Observable} from 'rxjs';
 import { AuthService } from '../services/auth.service';
 
 @Injectable({
@@ -16,9 +16,9 @@ export class AuthGuard implements CanActivate {
   canActivate(
     next: ActivatedRouteSnapshot,
     state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
-    const token = this.authService.tokenValue;
-    if (token) { return true; }
-    this.router.navigate(['/login']).then();
-    return false;
+    return true;
+    // if (token) { return true; }
+    // this.router.navigate(['/login']).then();
+    // return false;
   }
 }
