@@ -1,18 +1,14 @@
 import { Injectable } from '@angular/core';
 import { HttpRequest, HttpHandler, HttpEvent, HttpInterceptor } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { AuthService } from '../services/auth.service';
 
 @Injectable()
 export class JwtInterceptor implements HttpInterceptor {
 
-    constructor(private authService: AuthService) { }
+    constructor() { }
 
     intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-        const token = this.authService.tokenValue;
-        if (token) {
-            request = request.clone({ withCredentials: true });
-        }
+        request = request.clone({ withCredentials: true });
         return next.handle(request);
     }
 }
