@@ -1,5 +1,5 @@
 from django.contrib import admin
-from storage.models import Graph, Data
+from storage.models import Graph, Data, Statistics
 from django.utils.crypto import get_random_string
 from django.contrib.auth.hashers import make_password
 
@@ -26,5 +26,14 @@ class DataAdmin(admin.ModelAdmin):
     get_graph.admin_order_field = 'graph__id'
 
 
+class StatisticsAdmin(admin.ModelAdmin):
+    list_display = ('a_mean', 'w_mean', 'median', 'dominant',
+                    'std_deviation', 'highest', 'lowest', 'graph')
+    search_fields = ['graph__name']
+    list_filter = ['graph__type']
+    ordering = ['id']
+
+
 admin.site.register(Graph, GraphAdmin)
 admin.site.register(Data, DataAdmin)
+admin.site.register(Statistics, StatisticsAdmin)
