@@ -1,7 +1,8 @@
-# Generated by Django 2.2.5 on 2019-12-06 14:46
+# Generated by Django 3.1.4 on 2021-01-09 01:25
 
 from django.db import migrations, models
 import django.db.models.deletion
+import django.utils.timezone
 
 
 class Migration(migrations.Migration):
@@ -16,9 +17,31 @@ class Migration(migrations.Migration):
             name='Graph',
             fields=[
                 ('id', models.AutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
-                ('graph_name', models.CharField(max_length=32, unique=True)),
-                ('type', models.CharField(max_length=16)),
+                ('name', models.CharField(max_length=32)),
+                ('sub_name', models.CharField(max_length=32)),
+                ('description', models.CharField(max_length=64)),
+                ('type', models.CharField(choices=[('Line', 'Line'), ('Pie', 'Pie'), ('Bar-V', 'Bar Vertical'), ('Bar-H', 'Bar Horizontal')], default='Line', max_length=16)),
+                ('x_label', models.CharField(max_length=16)),
+                ('y_label', models.CharField(max_length=16)),
+                ('creation_date', models.DateTimeField(default=django.utils.timezone.now)),
             ],
+        ),
+        migrations.CreateModel(
+            name='Statistics',
+            fields=[
+                ('id', models.AutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
+                ('a_mean', models.FloatField(null=True)),
+                ('w_mean', models.FloatField(null=True)),
+                ('median', models.FloatField(null=True)),
+                ('dominant', models.FloatField(null=True)),
+                ('std_deviation', models.FloatField(null=True)),
+                ('highest', models.FloatField(null=True)),
+                ('lowest', models.FloatField(null=True)),
+                ('graph', models.OneToOneField(null=True, on_delete=django.db.models.deletion.CASCADE, related_name='statistics', to='storage.graph')),
+            ],
+            options={
+                'verbose_name_plural': 'Statistics',
+            },
         ),
         migrations.CreateModel(
             name='Data',
@@ -26,8 +49,7 @@ class Migration(migrations.Migration):
                 ('id', models.AutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
                 ('key', models.FloatField()),
                 ('value', models.FloatField()),
-                ('description', models.CharField(max_length=64)),
-                ('graph', models.ForeignKey(on_delete=django.db.models.deletion.CASCADE, to='storage.Graph')),
+                ('graph', models.ForeignKey(on_delete=django.db.models.deletion.CASCADE, related_name='data', to='storage.graph')),
             ],
             options={
                 'verbose_name_plural': 'Data',
